# api/gen/http/api/server/server.go
// Code generated by goa v3.2.3, DO NOT EDIT.
//
// api HTTP server
//
// Command:
// $ goa gen github.com/sm43/oauth/api/design

package server

import (
	"context"
	"net/http"

	api "github.com/sm43/oauth/api/gen/api"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
	"goa.design/plugins/v3/cors"
)

// Server lists the api service endpoint HTTP handlers.
type Server struct {
	Mounts       []*MountPoint
	Authenticate http.Handler
	Details      http.Handler
	CORS         http.Handler
}

// ErrorNamer is an interface implemented by generated error structs that
// exposes the name of the error as defined in the design.
type ErrorNamer interface {
	ErrorName() string
}

// MountPoint holds information about the mounted endpoints.
type MountPoint struct {
	// Method is the name of the service method served by the mounted HTTP handler.
	Method string
	// Verb is the HTTP method used to match requests to the mounted handler.
	Verb string
	// Pattern is the HTTP request path pattern used to match requests to the
	// mounted handler.
	Pattern string
}

// New instantiates HTTP handlers for all the api service endpoints using the
// provided encoder and decoder. The handlers are mounted on the given mux
// using the HTTP verb and path defined in the design. errhandler is called
// whenever a response fails to be encoded. formatter is used to format errors
// returned by the service methods prior to encoding. Both errhandler and
// formatter are optional and can be nil.
func New(
	e *api.Endpoints,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) *Server {
	return &Server{
		Mounts: []*MountPoint{
			{"Authenticate", "POST", "/oauth/redirect"},
			{"Details", "GET", "/details"},
			{"CORS", "OPTIONS", "/oauth/redirect"},
			{"CORS", "OPTIONS", "/details"},
		},
		Authenticate: NewAuthenticateHandler(e.Authenticate, mux, decoder, encoder, errhandler, formatter),
		Details:      NewDetailsHandler(e.Details, mux, decoder, encoder, errhandler, formatter),
		CORS:         NewCORSHandler(),
	}
}

// Service returns the name of the service served.
func (s *Server) Service() string { return "api" }

// Use wraps the server handlers with the given middleware.
func (s *Server) Use(m func(http.Handler) http.Handler) {
	s.Authenticate = m(s.Authenticate)
	s.Details = m(s.Details)
	s.CORS = m(s.CORS)
}

// Mount configures the mux to serve the api endpoints.
func Mount(mux goahttp.Muxer, h *Server) {
	MountAuthenticateHandler(mux, h.Authenticate)
	MountDetailsHandler(mux, h.Details)
	MountCORSHandler(mux, h.CORS)
}

// MountAuthenticateHandler configures the mux to serve the "api" service
// "Authenticate" endpoint.
func MountAuthenticateHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := handleAPIOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("POST", "/oauth/redirect", f)
}

// NewAuthenticateHandler creates a HTTP handler which loads the HTTP request
// and calls the "api" service "Authenticate" endpoint.
func NewAuthenticateHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeAuthenticateRequest(mux, decoder)
		encodeResponse = EncodeAuthenticateResponse(encoder)
		encodeError    = EncodeAuthenticateError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "Authenticate")
		ctx = context.WithValue(ctx, goa.ServiceKey, "api")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountDetailsHandler configures the mux to serve the "api" service "Details"
// endpoint.
func MountDetailsHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := handleAPIOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("GET", "/details", f)
}

// NewDetailsHandler creates a HTTP handler which loads the HTTP request and
// calls the "api" service "Details" endpoint.
func NewDetailsHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeDetailsRequest(mux, decoder)
		encodeResponse = EncodeDetailsResponse(encoder)
		encodeError    = EncodeDetailsError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "Details")
		ctx = context.WithValue(ctx, goa.ServiceKey, "api")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountCORSHandler configures the mux to serve the CORS endpoints for the
// service api.
func MountCORSHandler(mux goahttp.Muxer, h http.Handler) {
	h = handleAPIOrigin(h)
	f, ok := h.(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("OPTIONS", "/oauth/redirect", f)
	mux.Handle("OPTIONS", "/details", f)
}

// NewCORSHandler creates a HTTP handler which returns a simple 200 response.
func NewCORSHandler() http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(200)
	})
}

// handleAPIOrigin applies the CORS response headers corresponding to the
// origin for the service api.
func handleAPIOrigin(h http.Handler) http.Handler {
	origHndlr := h.(http.HandlerFunc)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		origin := r.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			origHndlr(w, r)
			return
		}
		if cors.MatchOrigin(origin, "*") {
			w.Header().Set("Access-Control-Allow-Origin", origin)
			w.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := r.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				w.Header().Set("Access-Control-Allow-Methods", "GET, POST")
				w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
			}
			origHndlr(w, r)
			return
		}
		origHndlr(w, r)
		return
	})
}