# api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: OAuth Demo
  description: HTTP services for OAuth API Demo
  version: "0.1"
servers:
- url: http://localhost:8000
paths:
  /details:
    get:
      tags:
      - api
      summary: Details api
      description: Find user details
      operationId: api#Details
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                githubID: sm43
                id: 1
                name: Shivam
        "401":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: true
                timeout: true
        "500":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: true
      security:
      - jwt_header_Authorization: []
  /oauth/redirect:
    post:
      tags:
      - api
      summary: Authenticate api
      description: Authenticates users against GitHub OAuth
      operationId: api#Authenticate
      parameters:
      - name: code
        in: query
        description: OAuth Authorization code of User
        allowEmptyValue: true
        required: true
        schema:
          type: string
          description: OAuth Authorization code of User
          example: Laboriosam expedita voluptas officia asperiores.
        example: Quia ut minus nemo iure.
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponseBody'
              example:
                token: Suscipit ab accusamus voluptatibus illum.
        "400":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: true
        "500":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: false
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: false
components:
  schemas:
    AuthenticateResponseBody:
      type: object
      properties:
        token:
          type: string
          description: JWT
          example: Suscipit similique laboriosam.
      example:
        token: Consectetur eveniet blanditiis provident explicabo.
      required:
      - token
    Error:
      type: object
      properties:
        fault:
          type: boolean
          description: Is the error a server-side fault?
          example: false
        id:
          type: string
          description: ID is a unique identifier for this particular occurrence of the problem.
          example: 123abc
        message:
          type: string
          description: Message is a human-readable explanation specific to this occurrence of the problem.
          example: parameter 'p' must be an integer
        name:
          type: string
          description: Name is the name of this class of errors.
          example: bad_request
        temporary:
          type: boolean
          description: Is the error temporary?
          example: true
        timeout:
          type: boolean
          description: Is the error a timeout?
          example: false
      description: Invalid Authorization code
      example:
        fault: true
        id: 123abc
        message: parameter 'p' must be an integer
        name: bad_request
        temporary: true
        timeout: true
      required:
      - name
      - id
      - message
      - temporary
      - timeout
      - fault
    User:
      type: object
      properties:
        githubID:
          type: string
          description: GitHub ID
          example: sm43
        id:
          type: integer
          description: ID is the unique id of User
          example: 1
        name:
          type: string
          description: Name of User
          example: Shivam
      example:
        githubID: sm43
        id: 1
        name: Shivam
      required:
      - id
      - name
      - githubID
  securitySchemes:
    jwt_header_Authorization:
      type: http
      description: Secures endpoint by requiring a valid JWT retrieved via the /oauth/redirect endpoint.
      scheme: bearer